backend/url_shortener/pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "celery>=5.6.3",
    "celery-types>=0.23.0",
    "django>=4.2.30",
    "django-celery-beat>=2.9.0",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.15.3",
    "flower>=2.0.1",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
]
